Added script to transfer a file from the client to the server
Created a Bash script that transfers a file from the client to the server using SSH and SCP. The script takes four command-line arguments: the path to the file to transfer, the server IP address, the server username, and the path to the SSH key file. The script checks for the required number of arguments and prompts the user if any are missing. It then uses SSH to create a directory on the server if it doesn't exist and uses SCP to transfer the file to the server. It also ensures that the SSH key is used with StrictHostKeyChecking disabled for convenience.