Added script to connect to the server using SSH and private key
Created a Bash script that establishes an SSH connection to the server at 54.234.54.214 using the private key located at ~/.ssh/school and the username 'ubuntu'. The script utilizes the 'ssh' command with the '-i' flag to specify the private key file and the hostname or IP address of the remote server.