Added script to generate an RSA key pair
Created a Bash script that generates an RSA key pair with a key length of 4096 bits and saves the private key to the file 'school' with the comment 'betty'. The script utilizes the 'ssh-keygen' command with the '-t' flag to specify the key type, the '-b' flag to specify the key length, the '-f' flag to specify the filename for the private key, and the '-C' flag to add a comment to the key pair.