Implement script to run whoami as another user
This script adds a new functionality to run the whoami command under the user specified as the first argument. It uses the su command with the -s /bin/bash option to switch to the target user's shell and then executes whoami, returning the username of the switched-to user. This script enables convenient identification of the current user within scripts or automation workflows involving multiple users.